requirements.txt
asgiref==3.4.1
Django==3.2.9
djangorestframework==3.12.4
mysqlclient==2.1.0
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.0
